Make the information update frequency a per client property
The demo html page now sends a number (the number of seconds between
updates) rather than a string. This number is now stored in the clients
data structure as update_freq and a new time_rem variable is also added
to the structure. This is initially set to the same value as
update_freq.

The timer now runs every second and looks through the active clients
decrementing the time_rem variable, when that hits 0 it sends the data
and resets the time_rem variable back to update_freq.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>