websocket_demo: Split up do_response()
Split out the actual gathering of data from do_response.

We now keep the hostname and network interface as global variables that
we setup once at startup.

We move the gathering of the uptime and network stats to their own
function get_stats() and now do_response simply calls this function,
builds the JSON and sends the data to the client.

This makes do_response easuer to read and reduces the amount of work
done when updating the client.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>