Allow to change the network interface to show stats for
Create a dropdown (using javascript) to provide a list of network
interfaces on the server with the default one selected.

It may not be particularly pretty but it seems to work.

We now store the selected network interface in the per client structure
as net_if.

We also now send an array of interface names in the JSON, this is
created on each server->client message and will allow for interfaces
coming and going, think virtual machine interfaces such as tap*.

A JSON message now looks like

    { "host": "server.example.com", "uptime": 47200, "ifname": "tap0",
     "rx": 16031, "tx": 45129, "ifnames": [ "lo", "enp3s0", "br0",
     "tap0" ] }

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>