Display the clients IP address
Display the IP address of the client on the web page. This comes in the
JSON in a new 'peerip' field. e.g

    { "host": "host.example.com", "peerip": "::1", "uptime": 10695,
     "ifname": "enp3s0", "rx": 188107926, "tx": 16054763, "ifnames":
     [ "lo", "enp3s0", "br0" ] }

Along with getpeername(2) this uses struct sockaddr_storage and
getnameinfo(3) to illustrate how to get the peers IP address as a
printable string in an address family agnostic manner.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>