websocket_demo: Add conditional support for libseccomp
libseccomp[0] provides an interface to the kernels seccomp system call
filtering interface. It's pretty neat.

This really just provides a simple demonstration of how it can be used.

Once websocket_demo has started up and bound to the listening sockets,
seccomp is setup to filter out all but the system calls actually needed
for further operation.

If libseccomp (the headers anyway at least, usually contained in a
-devel package) isn't installed, websocket_demo will fall back to being
built without this support.

[0] - https://github.com/seccomp/libseccomp

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>