websocket_demo.c: Re-work read_client_data()
This didn't handle the case where the client would send less than a
frames worth (2) of bytes and we'd get random values in the frame
header.

We also shouldn't call decode_frame() in a loop, just once we have at
least 2 bytes of data.

When reading from the client we should also handle receiving either of
EINTR, EAGAIN or EWOULDBLOCK.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>