websocket_demo: Fix byte order of 64bit values
The pay_len field of the websocket header defines the length of the data
sent. If it's <= 125 then that is the length, if it's == 126 then the
next 2 bytes taken as an unsigned int define the length and if it's ==
127 then the next 8 bytes taken as an unsigned int define the length.

Now the spec[0] says

    "Multibyte length quantities are expressed in network byte order."

while we took that into account for the 16bit pay_len case we didn't for
the 64bit case.

Use htobe64(3) & be64toh(3) to handle this.

[0] - https://tools.ietf.org/html/rfc6455#section-5.2

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>