Implement WebSockets over TLS (wss://)
This required some code restructuring especially to handle the TLS
handshake.

Some other bits got re-worked/removed for an overall improvement.

While this adds a dependency on OpenSSL it removes the use the GLib.

The use of TLS is disabled by default as it requires a bit of effort to
get going messing around with certificates etc. See the README for
details.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>