Avoid pointer arithmetic on void *
In net_read() & net_write() we pass the buffer in as a void * where we
then do pointer arithmetic on it.

While GCC does allow this by treating the size of void as 1[0]. It goes
against the C standard where it is undefined behaviour.

To avoid any possible confusions with people, tooling etc lets just use
char * for the buffer parameter instead.

[0]: https://gcc.gnu.org/onlinedocs/gcc/Pointer-Arith.html

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>